ocr_services/requirements.txt
# ═══════════════════════════════════════════════════════════════════
# OCR SERVICES - PRODUCTION REQUIREMENTS
# ═══════════════════════════════════════════════════════════════════
# Cài đặt: pip install -r requirements.txt
#
# ⚠️  MANUAL INSTALLATIONS (không có trong pip):
#   - torch, torchvision, torchaudio (cu118)
#   - vllm>=0.8.5+cu118
#   - flash-attn>=2.7.3
#   - transformers, tokenizers (quản lý bởi vllm)
#
# System dependencies:
#   - CUDA 11.8+
#   - cuDNN 8.x
#   - tesseract-ocr (cho fallback OCR)
# ═══════════════════════════════════════════════════════════════════

# ═══════════════════════════════════════════════════════════════════
# [1] WEB FRAMEWORK & API
# ═══════════════════════════════════════════════════════════════════
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# ═══════════════════════════════════════════════════════════════════
# [2] DATABASE & ORM
# ═══════════════════════════════════════════════════════════════════
sqlalchemy==2.0.24
psycopg2-binary==2.9.9
alembic==1.13.1

# ═══════════════════════════════════════════════════════════════════
# [3] MESSAGE QUEUE & ASYNC
# ═══════════════════════════════════════════════════════════════════
pika==1.3.2
celery==5.3.4
redis==5.0.1
kombu==5.3.4

# ═══════════════════════════════════════════════════════════════════
# [4] OBJECT STORAGE (MinIO/S3)
# ═══════════════════════════════════════════════════════════════════
minio==7.2.0
boto3==1.29.7
botocore==1.32.7
s3transfer==0.7.0

# ═══════════════════════════════════════════════════════════════════
# [5] PDF & IMAGE PROCESSING
# ═══════════════════════════════════════════════════════════════════
PyMuPDF==1.23.8              # fitz - PDF to images
img2pdf==1.4.1               # PIL images to PDF
Pillow==10.1.0               # Image processing (PIL)
opencv-python-headless==4.8.1.78  # deskew + image processing

# ═══════════════════════════════════════════════════════════════════
# [6] DEEP LEARNING & MODEL DEPENDENCIES
# ═══════════════════════════════════════════════════════════════════
# NOTE: Các packages này được quản lý bởi vllm, nhưng liệt kê để rõ ràng
einops==0.7.0
easydict==1.13
addict==2.4.0
accelerate==0.25.0
safetensors==0.4.1

# ═══════════════════════════════════════════════════════════════════
# [7] DATA PROCESSING & UTILITIES
# ═══════════════════════════════════════════════════════════════════
numpy==1.24.3
scipy==1.11.4
scikit-image==0.22.0
python-dotenv==1.0.0
tqdm==4.66.1
requests==2.31.0

# ═══════════════════════════════════════════════════════════════════
# [8] TESTING & DEVELOPMENT
# ═══════════════════════════════════════════════════════════════════
pytest==7.4.3
pytest-asyncio==0.21.1

# ═══════════════════════════════════════════════════════════════════
# [9] LOGGING & MONITORING
# ═══════════════════════════════════════════════════════════════════
python-json-logger==2.0.7

# ═══════════════════════════════════════════════════════════════════
# [10] TESSERACT OCR (Fallback engine) + IMAGE PROCESSING
# ═══════════════════════════════════════════════════════════════════
# Cần cài hệ thống: apt-get install tesseract-ocr tesseract-ocr-vie
pytesseract==0.3.10
imutils==0.5.4

# ═══════════════════════════════════════════════════════════════════
# [OPTIONAL] Các engines khác - uncomment nếu sử dụng
# ═══════════════════════════════════════════════════════════════════
# MinerU dependencies (uncomment nếu dùng MinerU engine)
# magic-pdf==0.5.0
# pdfminer.six==20221105
# detectron2==0.6

# Docling V2 (uncomment nếu dùng Docling engine)
# docling==0.1.0
# docling-core==0.1.0
